Move interfaces and options to Core
